requirements.txt
# Core
mcp
autogen-agentchat
autogen-ext[azure,mcp]
openai>=1.12.0
pyautogen[teachable]==0.2.9
flaml[automl]>=1.8.0
requests>=2.31.0
python-dotenv>=0.19.0
gitpython
python-dateutil
pydantic
PyYAML
fastapi 
uvicorn

# Azure SDK (demo 02 & 03)
azure-identity
azure-mgmt-resourcegraph
azure-mgmt-security
azure-mgmt-policyinsights   # ← solo si cambias REST→SDK

# GitHub SDK (opcional, demo 01)
PyGithub                    # si prefieres SDK
